Finished code examples and exercises for Context.
